Admin can now ban any user from posting and creating new recipes. Admin can also remove the ban if he wants it. In the list of all users admin can see banned icon next to the usernames.
on users profile there is also icon and text next to it so that both user and admin know that it is about ban. When clicked on icon and text admin can change the state of banned.

Banned users are disabled from clicking on new recipe button and also from manualy typing it in url.